Added validation of input file in inspect.py
inspect now verifies that the input is a REDCap data dictionary before proceeding; inspect exits and gives message if it is not.

created file _csv.py in djredcap/ to be used for reading csvs with correct encoding. also changed the encoding of test files to match the new (correct) output.

renamed djredcap subcommand to djconvert in order to avoid importing contradictions with its parent directory, djredcap.

Signed-off-by: James Swanick <swanijam@gmail.com>