Models.py now prints to stdout or given file.
Added -m or --model-file option in models.py, inspect.py, convert.py which
accepts a filename to which models are to be printed. If no filename is given,
models are printed to stdout.

Given filename is relative to the current working directory.

Tests.py has been updated to reflect the changes.

Signed-off-by: James Swanick <swanijam@gmail.com>